client/requirements.txt
dotenv==0.9.9
python-dotenv==1.2.1
websocket-client==1.9.0
